isar-bootstrap: Fix PREMIRRORS setting for cached build
First, there is no reason to match that closely on the repos used for
downloading Debian. That only breaks when using a different distro or
source.

Second, we must enforce the usage of base-apt in case of cached build.
Using a weak assignment will fail if the user specified
DISTRO_APT_PREMIRRORS for online build.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>