wic-img: Satisfy the quirks of latest wic
The latest WIC doesnot use fakeroot if pseudo_dir(IMAGE_ROOTFS/../pseudo)
is non-existent. This results in wic_fakeroot not being used. Create a
dummy pseudo_dir to satisfy wic.

WIC hardcodes the python(/usr/bin/python3-native/python3) to be used
while running bmaptool. The above path is OE specific and hence breaks
build in ISAR. Create a symbolic link in the above location to satisfy
wic.

In case of stretch, bmaptool is based on python2. In case of buster
and beyond it is python3. Grep the first line of bmaptool to find out
which version of python is used and symlink the appropriate binary
to /usr/bin/python3-native/python3.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>