wic_fakeroot: Handle standalone pseudo invocations
When using --exclude-path option wic copies the rootfs to a new location
and invokes pseudo as a standalone command to rebuild the database in the
new rootfs.

This is not applicable when using wic_fakeroot. Return 0 for such
standalone invocations in wic_fakeroot.

It also looks for files.db inside the pseudo directory and throws an
exception if it is not found. Handle that too.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>