meta-isar/canned-wks: Remove unwanted /boot mountpoint
In the older version of wic, the fstab entry for /boot mountpoint
was skipped. However in the latest wic this is not the case.
Adding a /boot mountpoint in part results in creating a fstab
entry for automount.

It is observed that in some kickstart files, /boot is used as a
mount point for bootloaders. This was okay with the previous
wic since anyway it ignored adding /boot mount points to fstab.

With the new wic, this now causes an unwanted mount of that partition
in /boot which inturn makes the original contents of /boot unavailable
after a successful boot.

This is not the default behaviour expected.
Remove such bogus /boot mount point in wks files.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>