meta-isar/canned-wks: Remove /boot mountpoint
In the older version of wic, the fstab entry for /boot mountpoint
was skipped.
However in the latest wic this is not the case. Adding a /boot
mountpoint in part results in creating a fstab entry for automount.
Remove it to avoid the unwanted mount which overwrites the contents
of /boot in debian rootfilesystem.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>