ci: Add compat arch support
Build samefile for the compat arch if that is enabled. Set
ISAR_ENABLE_COMPAT_ARCH for all supported combinations in CI.

As deploy this 32-bit package from the 64-bit image build would conflict
with the same package generated by the 32-bit image build, rename the
compat package in that case.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>