Fix legacy bios boot partition
In case of legacy bios, the kernel is required to be present
in a seperate partition with necessary syslinux files. This was
never mounted onto the target and standard kernel update via
apt-get was impossible.

Rename the existing kernel and initrd files in the syslinux
partition to the standard debian naming format. Copy over
System.map and config files to the partition and exclude the
/boot in root partition. The latest wic automatically adds the
necessary entry for /boot partition in fstab which automounts
it during boot.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>