meta-isar/canned-wks: Remove unwanted /boot mountpoint
In the older version of wic, the fstab entry for /boot mountpoint
was skipped. However in the latest wic this is not the case.
Adding a /boot mountpoint in part results in creating a fstab
entry for automount.

For machines that uses the kernel, initrd and related files from the root
partition; having /boot now as mountpoint for bootloader partition would
break the kernel update. Those original kernel files would no longer be
available once the bootloader partition is auto-mounted onto /boot.

For such machines, remove the bogus /boot mount point in wks files.

Signed-off-by: Vijai Kumar K <Vijaikumar_Kanagarajan@mentor.com>